tools/kwork-monitor/pyproject.toml
[project]
name = "kwork-monitor"
version = "0.1.0"
description = "Kwork freelance project monitor → Telegram alerts with AI evaluation"
requires-python = ">=3.12"
dependencies = [
    "kwork>=0.2.0",
    "httpx>=0.27",
    "google-genai>=1.0",
    "beartype>=0.19",
    "playwright>=1.58.0",
]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
